+++ src/components/abstract-component.js
@@ -1,0 +1,26 @@
+import {createElement} from "../utils/render.js";
+
+export default class AbstractComponent {
+  constructor() {
+    if (new.target === AbstractComponent) {
+      throw new Error(`Can't instantiate AbstractComponent, only concrete one.`);
+    }
+    this._element = null;
+  }
+  getTemplate() {
+    throw new Error(`Abstract method not implemented: getTemplate.`);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+
+}
