+++ src/components/sorting.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createSortTemplate = () => {
   return (
@@ -29,25 +29,8 @@ const createSortTemplate = () => {
   );
 };
 
-export default class Sorting {
-  constructor() {
-    this._element = null;
-
-  }
-
+export default class Sorting extends AbstractComponent{
   getTemplate() {
     return createSortTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
